Add users management pages on /users:
New users main page with search, filters and a table (similar to datasets main page).
Flags now configurable in the component FilteringView
which is used on main pages of users and datasets.

New single user page with tabs (similar to dataset view page, SingleDataView).
New component TabsView used on the view pages of user and dataset.
The single user page has two tabs:
 - "Details" which is form-formated with a button to submit changes.
 - "Logs" to view the logs of user management jobs.
New component MultiSelect used on the form of user details and on BodyEnumSelect.